move cache to lib
